refactor(dashboard): move dashboard_shell into apps/dashboard/client/layout
Per the framework's "app owns its routes" convention, the dashboard
app now owns its layout natively at `apps/dashboard/client/layout.rs`
instead of pulling `dashboard_shell` from `crate::client::layout` via
the `apps/dashboard/views.rs` re-export bridge.

Removes:
- `dashboard/src/client/layout.rs` (moved into the dashboard app).
- `dashboard/src/apps/dashboard/views.rs` (the only re-export bridge,
  no longer needed since `urls.rs` imports `dashboard_shell` directly
  from the new location and nothing else consumed
  `crate::apps::dashboard::views::not_found_page`).

Updates:
- `dashboard/src/apps/dashboard.rs` declares unconditional
  `pub mod client;` (mirroring the auth app pattern at
  `apps/auth.rs:14`); `pub mod views;` is dropped.
- `dashboard/src/apps/dashboard/urls.rs` imports `dashboard_shell` from
  `crate::apps::dashboard::client::layout`.
- `dashboard/src/client.rs` removes `pub mod layout;`.
- `dashboard/src/client/router.rs` (the `kent8192/reinhardt-web#4230`
  workaround) imports `dashboard_shell` from the new path; the
  workaround file itself is still tracked for removal in Phase E
  (`#577`).

Refs #578.
