+++ dashboard/src/apps/dashboard.rs
@@ -1,13 +1,17 @@
 //! Dashboard application module.
 //!
-//! Owns the project-level SPA route table — the home/clusters/deployments
-//! pages reachable from the top-level navigation. Views are pulled from
-//! `crate::client` until each section grows its own per-app implementation.
+//! Owns the project-level SPA `home` route. The dashboard shell layout
+//! (rendered for `/`) lives under `client/layout.rs`; per-section pages
+//! (`clusters:list`, `deployments:list`) belong to their respective
+//! apps.
 
 use reinhardt::app_config;
 
+// `client` is intentionally cross-target so the layout constructor is
+// reachable from the wasm SPA call sites and from the
+// `UnifiedRouter::client(...)` registration in `urls.rs`.
+pub mod client;
 pub mod urls;
-pub mod views;
 
 #[app_config(name = "dashboard", label = "dashboard")]
 pub struct DashboardConfig;
