+++ dashboard/src/apps/dashboard/client.rs
@@ -1,0 +1,6 @@
+//! WASM client modules for the dashboard application.
+//!
+//! Owns the dashboard shell layout used by the project-level `home`
+//! route, which is the primary entry point for the SPA.
+
+pub mod layout;
